refactor: Extract Art elements
